Errors
Error in eval(ei, envir) : attempt to apply non-function.
Causes
Misuse of Parentheses: Ensure that you are using parentheses () correctly when calling functions. If you use parentheses after a variable that is not a function, it will result in this error.
Variable Shadowing: Make sure you haven't accidentally assigned a value to a variable with the same name as a function you intend to use. This can cause confusion.
Incorrect Function Name: Ensure that the function you are trying to call actually exists and is spelled correctly.
Missing Function Definition: If you're encountering this error when defining a function, make sure you have used the correct syntax for function definitions.